Split MachineIdentityProviderTests into separate files, and used partials to group them under a single parent in the IDE.
